# docdoku-server/docdoku-server-web/src/main/webapp/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v2.5",
    "title" : "DocdokuPLM API"
  },
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/workspaces/{workspaceId}/folders/{folderId}/documents" : {
      "get" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Get documents from given params",
        "description" : "",
        "operationId" : "getDocuments",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "tagId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "query",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "assignedUserLogin",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "checkoutUser",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "filter",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "configSpec",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentRevisionDTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Create document",
        "description" : "",
        "operationId" : "createDocumentMasterInFolder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folderId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "configSpec",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/folders/{folderId}/documents/checkedout" : {
      "get" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Get checked out documents",
        "description" : "",
        "operationId" : "getCheckedOutDocMs",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentRevisionDTO"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{folderId}/documents/count" : {
      "get" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Count documents",
        "description" : "",
        "operationId" : "getDocumentsInWorkspaceCount",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CountDTO"
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{folderId}/documents/countCheckedOut" : {
      "get" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Count checked out documents",
        "description" : "",
        "operationId" : "countCheckedOutDocs",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentRevisionDTO"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{folderId}/documents/doc_revs" : {
      "get" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "searchDocumentRevisionsToLink : todo doc",
        "description" : "",
        "operationId" : "searchDocumentRevisionsToLink",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "l",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentRevisionDTO"
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}/folders/{folderId}/documents/search" : {
      "get" : {
        "tags" : [ "folders", "documents", "workspaces" ],
        "summary" : "Search documents",
        "description" : "",
        "operationId" : "searchDocumentRevision",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "workspaceId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentRevisionDTO"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ACLDTO" : {
      "type" : "object",
      "properties" : {
        "userEntries" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string",
            "enum" : [ "FORBIDDEN", "READ_ONLY", "FULL_ACCESS" ]
          }
        },
        "groupEntries" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string",
            "enum" : [ "FORBIDDEN", "READ_ONLY", "FULL_ACCESS" ]
          }
        }
      }
    },
...
    "CountDTO" : {
      "type" : "object",
      "properties" : {
        "count" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "DocumentIterationDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "documentMasterId" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "iteration" : {
          "type" : "integer",
          "format" : "int32"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "modificationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "checkInDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "title" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "revisionNote" : {
          "type" : "string"
        },
        "attachedFiles" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "instanceAttributes" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InstanceAttributeDTO"
          }
        },
        "linkedDocuments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DocumentRevisionDTO"
          }
        },
        "documentRevision" : {
          "$ref" : "#/definitions/DocumentRevisionDTO"
        }
      }
    },
...
    "DocumentRevisionDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "id" : {
          "type" : "string"
        },
        "documentMasterId" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "author" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "commentLink" : {
          "type" : "string"
        },
        "title" : {
          "type" : "string"
        },
        "checkOutUser" : {
          "$ref" : "#/definitions/UserDTO"
        },
        "checkOutDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "description" : {
          "type" : "string"
        },
        "iterationSubscription" : {
          "type" : "boolean",
          "default" : false
        },
        "stateSubscription" : {
          "type" : "boolean",
          "default" : false
        },
        "documentIterations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/DocumentIterationDTO"
          }
        },
        "workflow" : {
          "$ref" : "#/definitions/WorkflowDTO"
        },
        "path" : {
          "type" : "string"
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "publicShared" : {
          "type" : "boolean",
          "default" : false
        },
        "attributesLocked" : {
          "type" : "boolean",
          "default" : false
        },
        "acl" : {
          "$ref" : "#/definitions/ACLDTO"
        },
        "lastIteration" : {
          "$ref" : "#/definitions/DocumentIterationDTO"
        }
      }
    },
...
    "InstanceAttributeDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "mandatory" : {
          "type" : "boolean",
          "default" : false
        },
        "locked" : {
          "type" : "boolean",
          "default" : false
        },
        "type" : {
          "type" : "string",
          "enum" : [ "TEXT", "NUMBER", "DATE", "BOOLEAN", "URL", "LOV" ]
        },
        "value" : {
          "type" : "string"
        },
        "lovName" : {
          "type" : "string"
        },
        "items" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/NameValuePairDTO"
          }
        }
      }
    },
...
    "NameValuePairDTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        }
      }
    },
...
    "UserDTO" : {
      "type" : "object",
      "properties" : {
        "workspaceId" : {
          "type" : "string"
        },
        "login" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "email" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        },
        "membership" : {
          "type" : "string",
          "enum" : [ "READ_ONLY", "FULL_ACCESS" ]
        }
      }
    },
...
    "WorkflowDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "finalLifeCycleState" : {
          "type" : "string"
        },
        "activities" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ActivityDTO"
          }
        },
        "abortedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lifeCycleState" : {
          "type" : "string"
        },
        "currentActivity" : {
          "$ref" : "#/definitions/ActivityDTO"
        },
        "currentStep" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
  }
}